requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
click==8.1.7
fastapi==0.115.4
h11==0.14.0
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
outcome==1.3.0.post0
pandas==2.2.3
phonenumbers==8.13.49
pillow==11.0.0
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websocket-client==1.8.0
wsproto==1.2.0
